backend/pyproject.toml
[project]
name = "dbf-comparator-pro-backend"
version = "0.1.0"
description = "Backend service for DBF Comparator Pro web version"
requires-python = ">=3.12"
dependencies = [
  "alembic>=1.16.0",
  "celery[redis]>=5.4.0",
  "dbfread>=2.0.7",
  "fastapi>=0.115.0",
  "psycopg[binary]>=3.2.0",
  "pydantic-settings>=2.8.0",
  "python-multipart>=0.0.20",
  "sqlalchemy>=2.0.40",
  "uvicorn[standard]>=0.34.0",
  "XlsxWriter>=3.2.0",
]

[project.optional-dependencies]
dev = [
  "httpx>=0.28.0",
  "pytest>=8.3.0",
  "pytest-asyncio>=0.26.0",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["app*"]

[tool.pytest.ini_options]
testpaths = ["tests"]

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"
